fix: Fix with label for single item & Refactor by adding edit-category
