+++ components/common/BottomSheet/SheetWithCourse.tsx
@@ -7,7 +7,7 @@ import {
   SheetTrigger,
 } from "@/components/ui/sheet";
 import Image from "next/image";
-import { OrderType2 } from "@/app/edit-course/_components/DragAndDropArea";
+import { OrderType2 } from "@/app/edit-category/_components/DragAndDropArea";
 import { iconInfo } from "@/lib/utils";
 
 interface SheetWithCourseProps {
